+++ src/auth/session/storage/__tests__/mysql.test.ts
@@ -14,7 +14,7 @@ const dbURL = new URL('mysql://shopify:passify@localhost/shopitest');
 
 // SORRY NOT SORRY. Docker containers can take quite a while to get ready,
 // especially on CI. This is hopefully enough.
-jest.setTimeout(20000);
+jest.setTimeout(25000);
 
 describe('MySQLSessionStorage', () => {
   let storage: MySQLSessionStorage;